Separate live send execution context contracts
