+++ src/PlateauResoniteLink/Targets/Resonite/ResoniteLiveSceneImportExecutionContext.cs
@@ -1,0 +1,7 @@
+namespace PlateauResoniteLink.Targets.Resonite;
+
+internal sealed record ResoniteLiveSceneImportExecutionContext(
+    ResoniteImportMemoryProfile MemoryProfile,
+    int ConnectionCount,
+    bool MeshBakeEnabled,
+    ResoniteLiveSendTargetContext LiveSendContext);
