+++ src/PlateauResoniteLink/Targets/Resonite/ResoniteLiveSceneImportExecutor.cs
@@ -91,9 +91,3 @@ internal sealed class ResoniteLiveSceneImportExecutor(
         }
     }
 }
-
-internal sealed record ResoniteLiveSceneImportExecutionContext(
-    ResoniteImportMemoryProfile MemoryProfile,
-    int ConnectionCount,
-    bool MeshBakeEnabled,
-    ResoniteLiveSendTargetContext LiveSendContext);
