+++ src/PlateauResoniteLink/Targets/Resonite/ResoniteLiveSendTargetContext.cs
@@ -1,0 +1,12 @@
+using System;
+
+using PlateauResoniteLink.Transport.ResoniteLink;
+
+namespace PlateauResoniteLink.Targets.Resonite;
+
+internal sealed record ResoniteLiveSendTargetContext(
+    Uri Endpoint,
+    int ConnectionCount,
+    ILiveSendClientSession ClientSession,
+    ResoniteLinkSendDiagnostics Diagnostics,
+    Action<string>? ProgressReporter);
